Rewrite PillLabel with TypeScript
